moved accounts config code to its own file
